+++ OSM POI Editor/OPEWay.h
@@ -27,8 +27,8 @@
 
 @property (nonatomic,strong) NSArray * nodes;
 
--(id)initWithArrayOfNodes:(NSArray *)arrayOfNodes tags:(NSMutableDictionary *)tagDictioanry ID: (int) i version:(int) version ;
--(id)initWithArrayOfNodes:(NSArray *)arrayOfNodes ID: (int) i version:(int) version;
+-(id)initWithArrayOfNodes:(NSArray *)arrayOfNodes tags:(NSMutableDictionary *)tagDictioanry ID: (int64_t) i version:(int) version ;
+-(id)initWithArrayOfNodes:(NSArray *)arrayOfNodes ID: (int64_t) i version:(int) version;
 
 -(void)prepareToDelete:(NSArray *)keys;
